+++ src/components/workbook/SheetPane.js
@@ -353,7 +353,7 @@ const SheetPane = props => {
           onClick={() => onMoveSheet(0)}
           className={`btn btn-sm btn-${
             theme === "dark" ? "dark" : "white"
-          } border-0 border-end px-2 rounded-0`}
+          } border-0 border-end border-secondary px-2 rounded-0`}
           disabled={sheets.findIndex(s => s.id === activeSheet) === 0}
         >
           <i className='fa fa-chevron-left' />
@@ -389,7 +389,7 @@ const SheetPane = props => {
                       </OverlayTrigger>
                       <div
                         style={{ minWidth: 130 }}
-                        className={`rounded-0 btn btn-sm btn-${
+                        className={`rounded-0 border-start btn btn-sm btn-${
                           activeSheet === sheet.id ? "bni" : theme
                         } border-0 border-end ${
                           theme === "dark" ? "border-secondary" : ""
@@ -409,7 +409,7 @@ const SheetPane = props => {
           onClick={() => onMoveSheet(1)}
           className={`btn btn-sm btn-${
             theme === "dark" ? "dark" : "white"
-          } border-0 border-0 border-start border-end px-2 rounded-0`}
+          } border-0 border-0 border-start border-end border-secondary px-2 rounded-0`}
           disabled={
             sheets.findIndex(s => s.id === activeSheet) === sheets.length - 1
           }