+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [2.8.7](https://github.com/bharani-palani/ledgerely/compare/v2.8.6...v2.8.7) (2025-05-14)
+
+
+### Bug Fixes
+
+* chart options changed to floating menu ([8d20ec9](https://github.com/bharani-palani/ledgerely/commit/8d20ec958d14f5777c9cdf0b03b0441a8b5d8313))
+
+## [2.8.6](https://github.com/bharani-palani/ledgerely/compare/v2.8.5...v2.8.6) (2025-05-09)
+
+
+### Bug Fixes
+
+* workbook charts moved to npm rnd ([5bce3be](https://github.com/bharani-palani/ledgerely/commit/5bce3bee65bf719f5b084c1fb1c968363caf2601))
+
 ## [2.8.5](https://github.com/bharani-palani/ledgerely/compare/v2.8.4...v2.8.5) (2025-05-07)